// packages/oslo-converter-uml-ea/lib/enums/TagName.ts
export enum TagName {
  ApCodelist = 'ap-codelist',
  ApDefinition = 'ap-definition',
  ApLabel = 'ap-label',
  ApUsageNote = 'ap-usageNote',
  AssociationDestPrefix = 'target-',
  AssociationDestRevPrefix = 'target-rev-',
  AssociationSourcePrefix = 'source-',
  AssociationSourceRevPrefix = 'source-rev-',
  DefiningPackage = 'package',
  Definition = 'definition',
  ExternalUri = 'uri',
  Ignore = 'ignore',
  IsLiteral = 'literal',
  Label = 'label',
  LocalName = 'name',
  PackageBaseUri = 'baseURI',
  PackageOntologyUri = 'ontologyURI',
  PackageBaseUriAbbreviation = 'baseURIabbrev',
  ParentUri = 'parentURI',
  Range = 'range',
  UsageNote = 'usageNote'
}